backend/ai-service/requirements.txt
# =============================================================================
# AI Service Dependencies
# =============================================================================

# --- Web Framework ---
fastapi==0.115.0
uvicorn[standard]==0.30.6

# --- AI / LLM ---
google-generativeai==0.8.3
langgraph==0.2.60
langchain-core==0.3.28

# --- LangSmith (Tracing & Testing) ---
langsmith==0.2.10

# --- Database ---
sqlalchemy==2.0.36

# --- Utilities ---
python-dotenv==1.0.1
pydantic==2.10.0
httpx==0.28.0

# --- CORS ---
# (included in fastapi)
